Change widget for logs Switch to the paragraph widget from termui for the log lines in the pod details view.
The list widget is broken if we use `WrapText = true` and the list
contains a log line which use multiple lines. In this case the scrolling
does not work properly.

An other option would be to use `WrapText = false` ut then some
important parts of a log line are cutted.